Split out storage and test utils into separate modules
