+++ CMakeLists.txt
@@ -124,6 +124,7 @@ elseif (FAASM_BUILD_TYPE STREQUAL "lambda-worker")
     add_subdirectory(src/redis)
     add_subdirectory(src/scheduler)
     add_subdirectory(src/state)
+    add_subdirectory(src/storage)
     add_subdirectory(src/system)
     add_subdirectory(src/util)
     add_subdirectory(src/wasm)
@@ -197,6 +198,7 @@ else ()
     add_subdirectory(src/runner)
     add_subdirectory(src/scheduler)
     add_subdirectory(src/state)
+    add_subdirectory(src/storage)
     add_subdirectory(src/system)
     add_subdirectory(src/upload)
     add_subdirectory(src/util)
@@ -206,5 +208,6 @@ else ()
     # Tests
     add_library(catch INTERFACE)
     add_subdirectory(test)
+    add_subdirectory(test_utils)
     add_subdirectory(test_integration)
 endif ()
