+++ src/storage/CMakeLists.txt
@@ -1,0 +1,18 @@
+include_directories(
+        ${FAASM_INCLUDE_DIR}/storage
+        ${FAASM_WAVM_SOURCE_DIR}
+)
+
+file(GLOB HEADERS "${FAASM_INCLUDE_DIR}/storage/*.h")
+
+set(LIB_FILES
+        FunctionLoader.cpp
+        LocalFunctionLoader.cpp
+        FileserverFunctionLoader.cpp
+        S3FunctionLoader.cpp
+        instance.cpp
+        ${HEADERS}
+        )
+
+faasm_private_lib(storage "${LIB_FILES}")
+target_link_libraries(storage util aws WAVM)
