+++ test/CMakeLists.txt
@@ -1,18 +1,16 @@
 file(GLOB_RECURSE TEST_FILES ${CMAKE_CURRENT_LIST_DIR} test_*.cpp)
 
-include_directories(${CMAKE_CURRENT_LIST_DIR})
+include_directories(
+        ${CMAKE_CURRENT_LIST_DIR}
+        ${CMAKE_CURRENT_LIST_DIR}/../test_utils
+)
 
 add_executable(
         tests
         main.cpp
-        utils/matrix_utils.cpp
-        utils/message_utils.cpp
-        utils/system_utils.cpp
-        utils/worker_utils.cpp
-        utils.h
         ${TEST_FILES}
 )
 
-target_link_libraries(tests catch faasm edge_lib upload_lib worker_lib emulator data)
+target_link_libraries(tests test_utils)
 
 add_test(NAME tests COMMAND "test/tests")
