--- test/utils/message_utils.cpp
@@ -1,21 +1,0 @@
-#include <catch/catch.hpp>
-
-#include "utils.h"
-
-using namespace faasm;
-
-namespace tests {
-    void checkMessageEquality(const message::Message &msgA, const message::Message &msgB) {
-        REQUIRE(msgA.user() == msgB.user());
-        REQUIRE(msgA.function() == msgB.function());
-        REQUIRE(msgA.isasync() == msgB.isasync());
-        REQUIRE(msgA.ispython() == msgB.ispython());
-        REQUIRE(msgA.success() == msgB.success());
-        REQUIRE(msgA.inputdata() == msgB.inputdata());
-        REQUIRE(msgA.outputdata() == msgB.outputdata());
-        REQUIRE(msgA.resultkey() == msgB.resultkey());
-        REQUIRE(msgA.statuskey() == msgB.statuskey());
-        REQUIRE(msgA.type() == msgB.type());
-    }
-}
-
