+++ test_integration/CMakeLists.txt
@@ -1,6 +1,9 @@
 file(GLOB_RECURSE TEST_FILES ${CMAKE_CURRENT_LIST_DIR} test_*.cpp)
 
-include_directories(${CMAKE_CURRENT_LIST_DIR})
+include_directories(
+        ${CMAKE_CURRENT_LIST_DIR}
+        ${CMAKE_CURRENT_LIST_DIR}/../test_utils
+)
 
 add_executable(
         integration_tests
@@ -8,6 +11,6 @@ add_executable(
         ${TEST_FILES}
 )
 
-target_link_libraries(integration_tests scheduler wasm aws)
+target_link_libraries(integration_tests test_utils scheduler wasm aws)
 
 add_test(NAME integration_tests COMMAND "integration_test/tests")
